+++ crates/filter-list-manager/src/manager/models/filter_list_rules_raw.rs
@@ -12,4 +12,7 @@ pub struct FilterListRulesRaw {
     pub rules: String,
     /// List of only disabled rules.
     pub disabled_rules: String,
+    /// Rules count in this filter list. Simply a number of non-empty lines
+    /// and does not start with a comment marker.
+    pub rules_count: i32,
 }