+++ platform/apple/AdGuardFLM/Sources/AdGuardFLMLib/GeneratedProto/filters.pb.swift
@@ -69,6 +69,10 @@ public struct FilterListManager_FilterListRules: Sendable {
   /// List of only disabled rules.
   public var disabledRules: [String] = []
 
+  //// Rules count in this filter list. Simply a number of non-empty lines
+  //// and does not start with a comment marker.
+  public var rulesCount: Int32 = 0
+
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
   public init() {}
@@ -87,6 +91,10 @@ public struct FilterListManager_FilterListRulesRaw: Sendable {
 
   /// List of only disabled rules as string.
   public var disabledRules: String = String()
+
+  //// Rules count in this filter list. Simply a number of non-empty lines
+  //// and does not start with a comment marker.
+  public var rulesCount: Int32 = 0
 
   public var unknownFields = SwiftProtobuf.UnknownStorage()
 
@@ -445,6 +453,23 @@ public struct FilterListManager_FullFilterList: @unchecked Sendable {
   fileprivate var _storage = _StorageClass.defaultInstance
 }
 
+public struct FilterListManager_RulesCountByFilter: Sendable {
+  // SwiftProtobuf.Message conformance is added in an extension below. See the
+  // `Message` and `Message+*Additions` files in the SwiftProtobuf library for
+  // methods supported on all messages.
+
+  /// Associated filter id.
+  public var filterID: Int32 = 0
+
+  /// Rules count in this filter list. Simply a number of non-empty lines
+  /// and does not start with a comment marker.
+  public var rulesCount: Int32 = 0
+
+  public var unknownFields = SwiftProtobuf.UnknownStorage()
+
+  public init() {}
+}
+
 // MARK: - Code below here is support for the SwiftProtobuf runtime.
 
 fileprivate let _protobuf_package = "filter_list_manager"
@@ -537,6 +562,7 @@ extension FilterListManager_FilterListRules: SwiftProtobuf.Message, SwiftProtobu
     1: .standard(proto: "filter_id"),
     2: .same(proto: "rules"),
     3: .standard(proto: "disabled_rules"),
+    4: .standard(proto: "rules_count"),
   ]
 
   public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -548,6 +574,7 @@ extension FilterListManager_FilterListRules: SwiftProtobuf.Message, SwiftProtobu
       case 1: try { try decoder.decodeSingularInt32Field(value: &self.filterID) }()
       case 2: try { try decoder.decodeRepeatedStringField(value: &self.rules) }()
       case 3: try { try decoder.decodeRepeatedStringField(value: &self.disabledRules) }()
+      case 4: try { try decoder.decodeSingularInt32Field(value: &self.rulesCount) }()
       default: break
       }
     }
@@ -563,6 +590,9 @@ extension FilterListManager_FilterListRules: SwiftProtobuf.Message, SwiftProtobu
     if !self.disabledRules.isEmpty {
       try visitor.visitRepeatedStringField(value: self.disabledRules, fieldNumber: 3)
     }
+    if self.rulesCount != 0 {
+      try visitor.visitSingularInt32Field(value: self.rulesCount, fieldNumber: 4)
+    }
     try unknownFields.traverse(visitor: &visitor)
   }
 
@@ -570,6 +600,7 @@ extension FilterListManager_FilterListRules: SwiftProtobuf.Message, SwiftProtobu
     if lhs.filterID != rhs.filterID {return false}
     if lhs.rules != rhs.rules {return false}
     if lhs.disabledRules != rhs.disabledRules {return false}
+    if lhs.rulesCount != rhs.rulesCount {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -581,6 +612,7 @@ extension FilterListManager_FilterListRulesRaw: SwiftProtobuf.Message, SwiftProt
     1: .standard(proto: "filter_id"),
     2: .same(proto: "rules"),
     3: .standard(proto: "disabled_rules"),
+    4: .standard(proto: "rules_count"),
   ]
 
   public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
@@ -592,6 +624,7 @@ extension FilterListManager_FilterListRulesRaw: SwiftProtobuf.Message, SwiftProt
       case 1: try { try decoder.decodeSingularInt32Field(value: &self.filterID) }()
       case 2: try { try decoder.decodeSingularStringField(value: &self.rules) }()
       case 3: try { try decoder.decodeSingularStringField(value: &self.disabledRules) }()
+      case 4: try { try decoder.decodeSingularInt32Field(value: &self.rulesCount) }()
       default: break
       }
     }
@@ -607,6 +640,9 @@ extension FilterListManager_FilterListRulesRaw: SwiftProtobuf.Message, SwiftProt
     if !self.disabledRules.isEmpty {
       try visitor.visitSingularStringField(value: self.disabledRules, fieldNumber: 3)
     }
+    if self.rulesCount != 0 {
+      try visitor.visitSingularInt32Field(value: self.rulesCount, fieldNumber: 4)
+    }
     try unknownFields.traverse(visitor: &visitor)
   }
 
@@ -614,6 +650,7 @@ extension FilterListManager_FilterListRulesRaw: SwiftProtobuf.Message, SwiftProt
     if lhs.filterID != rhs.filterID {return false}
     if lhs.rules != rhs.rules {return false}
     if lhs.disabledRules != rhs.disabledRules {return false}
+    if lhs.rulesCount != rhs.rulesCount {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }
@@ -1234,6 +1271,44 @@ extension FilterListManager_FullFilterList: SwiftProtobuf.Message, SwiftProtobuf
       }
       if !storagesAreEqual {return false}
     }
+    if lhs.unknownFields != rhs.unknownFields {return false}
+    return true
+  }
+}
+
+extension FilterListManager_RulesCountByFilter: SwiftProtobuf.Message, SwiftProtobuf._MessageImplementationBase, SwiftProtobuf._ProtoNameProviding {
+  public static let protoMessageName: String = _protobuf_package + ".RulesCountByFilter"
+  public static let _protobuf_nameMap: SwiftProtobuf._NameMap = [
+    1: .standard(proto: "filter_id"),
+    2: .standard(proto: "rules_count"),
+  ]
+
+  public mutating func decodeMessage<D: SwiftProtobuf.Decoder>(decoder: inout D) throws {
+    while let fieldNumber = try decoder.nextFieldNumber() {
+      // The use of inline closures is to circumvent an issue where the compiler
+      // allocates stack space for every case branch when no optimizations are
+      // enabled. https://github.com/apple/swift-protobuf/issues/1034
+      switch fieldNumber {
+      case 1: try { try decoder.decodeSingularInt32Field(value: &self.filterID) }()
+      case 2: try { try decoder.decodeSingularInt32Field(value: &self.rulesCount) }()
+      default: break
+      }
+    }
+  }
+
+  public func traverse<V: SwiftProtobuf.Visitor>(visitor: inout V) throws {
+    if self.filterID != 0 {
+      try visitor.visitSingularInt32Field(value: self.filterID, fieldNumber: 1)
+    }
+    if self.rulesCount != 0 {
+      try visitor.visitSingularInt32Field(value: self.rulesCount, fieldNumber: 2)
+    }
+    try unknownFields.traverse(visitor: &visitor)
+  }
+
+  public static func ==(lhs: FilterListManager_RulesCountByFilter, rhs: FilterListManager_RulesCountByFilter) -> Bool {
+    if lhs.filterID != rhs.filterID {return false}
+    if lhs.rulesCount != rhs.rulesCount {return false}
     if lhs.unknownFields != rhs.unknownFields {return false}
     return true
   }