+++ platform/apple/AdGuardFLM/Tests/AdGuardFLMLibTests/AdGuardFLMLibTests.swift
@@ -96,6 +96,9 @@ final class AdGuardFLMLibTests: XCTestCase {
         let disabledRules = try flm.getDisabledRules(ids: [customFilterFromString.id])
         NSLog("GetDisabledRules returns \(disabledRules)")
 
+        let rulesCount = try flm.getRulesCount(ids: [customFilterFromString.id])
+        NSLog("GetRulesCount return \(rulesCount)")
+
         XCTAssertNoThrow(try flm.deleteCustomFilterLists(ids: [customFilterFromString.id]))
 
         NSLog("Filter \(customFilterFromString.id) was deleted")
@@ -150,6 +153,7 @@ final class AdGuardFLMLibTests: XCTestCase {
         XCTAssert(constants.special_group_id == 0, "FLM Special group id must be zero")
         XCTAssert(constants.smallest_filter_id == -2_000_000_000, "FLM Custom group id must be two billions")
         XCTAssert(filterMetadataWithBody.metadata.homepage.count > 0, "FLM Metadata homepage must be non-empty")
+        XCTAssert(rulesCount[0].rulesCount == 2, "FLM Rules count must be 2")
     }
 }